Added more pid

+++ inc/screen.h
@@ -1,4 +1,4 @@
 //screen.h
 void adc_init(void);
-uint32_t screen_getX(void);
-uint32_t screen_getY(void);
+float screen_getX(void);
+float screen_getY(void);
